requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
attrs
beautifulsoup4
bs4
cachetools
certifi
cffi
charset-normalizer
click
colorama
discord.py
fastapi
frozenlist
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
googleapis-common-protos
h11
httpcore
httplib2
httptools
httpx
idna
itsdangerous
Jinja2
MarkupSafe
multidict
oauthlib
propcache
proto-plus
protobuf
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic_core
PyNaCl
pyparsing
pypresence
python-dotenv
pytz
PyYAML
requests
requests-oauthlib
rsa
sniffio
soupsieve
starlette
typing-inspection
typing_extensions
uritemplate
urllib3
uvicorn
watchfiles
websockets
yarl
youtube-dl
yt-dlp